Enable stricter version of Naming/FileName cop
The `ExpectMatchingDefinition` option for this cop is `false` by
default. However, setting it to `true` is useful because it identifies
things like typos in names of classes, modules and their files.

I've made a few exceptions, because they weren't straitforward to fix,
but it would be nice to avoid these exceptions at some point.