Remove argument splat in Deprecation.warning
There's only one callsite for this method and it doesn't make use of the
splatting. I considered changing that caller to make use of the
splatting, but I decided against it in the end, because in that case we
want to join the messages with a space character and that might not
always be the case.