Move deprecation assertions outside inner test
Previously I think it was confusing that the deprecation warnings
assertions were inside the inner tests in the
`KeywordArgumentMatchingTest` acceptance test. Now the assertions are
done outside the inner tests. The latter means that we also need to
declare the `execution_point` local variable outside the inner tests.

The above change means we no longer need to include the
`DeprecationCapture` module into the outer test.

Adding the `deprecation_warnings` accessor to the instance of
`Test::Unit::TestResult` built by `TestUnitResult` is probably overkill,
because I believe the latter is only used in the integration tests which
don't (currently) need access to the captured deprecation warnings.
However, I decided it was less confusing to add it for now.