separate rubocop configuration for tests
As stated in commits e423977, 33e373c, fddebc3:

Ideally I would've configured the extra `AllowedMethods` to only apply
to the tests, but I couldn't find a simple way to do that with the
rubocop configuration. It might be worth extracting a separate rubocop
configuration for the tests.

Here we do just that.