Fix compatibility with ruby 3.5
In 3.5, most of the `cgi` gem is removed. Only methods relating to escaping/unescaping are retained.

That means, on 3.5 you get this error:
> NoMethodError: undefined method 'parse' for class CGI

This replaces the cgi usage with https://docs.ruby-lang.org/en/master/URI.html#method-c-decode_www_form, that also returns this information.

It's in array form, so it first needs to be converted to an hash.

This method is available on all versions going back to at least Ruby 2.0 so compatibility should be ok with this.

https://bugs.ruby-lang.org/issues/21258