Add assertion count tests for unexpected invocation
This new test documents some existing correct behaviour, i.e. when an
unexpected method is invoked, the test fails fast without incrementing
the assertion count as long as there was no explicit "never"
expectation for the invoked method.

Note that this is distinct from the scenario identified in #763 where
there *is* an explicit "never" expectation for the invoked method.