Never expectation should increment assertion count
When a test includes an explicit "never" expectation for a method and
that method is called, the assertion count should be incremented.

Since activesupport v7.2.0 [1] and specifically this PR [2], Rails has
displayed a warning for tests with no assertions. Previously, a test
where the only assertion was a Mocha "never" expectation would have
incorrectly displayed such a warning.

Note that we're now somewhat unnecessarily passing the
`AssertionCounter` into both `Hooks#mocha_setup` and
`Hooks#mocha_verify`. I plan to rationalize the code in subsequent
commits.

Fixes #763.

[1]: https://github.com/rails/rails/blob/v7.2.0/activesupport/CHANGELOG.md
[2]: https://github.com/rails/rails/pull/51625