ParameterMatchers#parse_option -> HasEntry.parse_option
This method is specific to the `HasEntry` matcher and so it belongs on
that class. Making it a class method rather than an instance method and
public rather than private will make some other changes I'm planning
easier. Given it's not included in the public API, there's no need to
provide a deprecation warning for this change.