Extract ParameterMatchers::Methods module
This moves the matcher builder methods previously defined in
`ParameterMatchers` into a new `ParameterMatchers::Methods` module.

It also changes `Mocha::API` so that it only includes the new
`ParameterMatchers::Methods` module and not the whole of
`ParameterMatchers`.

My main motivation for this change is to address #766 by only including
the matcher builder methods (e.g. `#has_entry`) and not the matcher
classes themselves (e.g. `HasEntry`) into the test/spec class in order
to reduce the "pollution" of its namespace.

Since the matcher builder methods and the matcher classes are part of
the public API, we'll need to generate deprecation warnings for this
change. I plan to address that in subsequent commits.

Although it's no longer possible to reference matcher classes from
within a test/spec using a "relative" constant lookup, e.g. `HasEntry`,
it's still possible to reference them by using an "absolute" constant
lookup, e.g.  `Mocha::ParameterMatchers::HasEntry`. This is how I fixed
a number of the tests in `HasEntryTest`.

I've updated a number of YARD documentation links to point to the new
location of the relevant method definitions.

The changes in this commit are best viewed using the
`--ignore-all-space` option on `git diff` to hide the
indentation-related changes.

Addresses #766.