Add deprecated ParameterMatchers matcher classes
In a previous commit, `Mocha::API` was changed so that it only includes
the new `ParameterMatchers::Methods` module and not the whole of
`ParameterMatchers`.

However, since these classes are part of the public API, we need to
retain the original versions of the classes and generate a deprecation
warning if they are used. For example, previously the following test
would have been able to access the `Anything` class using "relative"
constant lookup from within a test/spec:

    class MyTest < Minitest::Test
      def test_me
        Anything.new
      end
    end

In order to keep this code working, this commit intercepts constant
lookup on the `ParameterMatchers::Methods` module (which is automatically
included into the test/spec) and returns the canonical matcher class
from the `ParameterMatchers` module while generating a deprecation
warning.

I've defined some shared tests in `ParameterMatchersTestHelper` for this
behaviour which are included into the individual matcher unit tests.