Improve matcher class deprecation warning location
Previously these deprecation warnings were all displaying
`<internal:kernel>:91:in 'Kernel#tap'` as the location, because
`Kernel#tap` is a Ruby built-in with a fake source location.

By removing the call to `Kernel#tap`, we can ensure the location is
given as the line in the test/spec which references a matcher class.